requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.1
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.4.0
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==15.0.7
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.28.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
triton==2.0.0
typing-extensions==4.4.0
urllib3==1.26.13
zipp==3.17.0
